requirements.txt
flask==1.1.2
gunicorn==20.0.4
logzero==1.5.0
python-telegram-bot==12.7
